Reorganize repo and update readme
